requirements.txt
apscheduler==3.11.1
flasgger==0.9.7.1
Flask==3.1.2
Flask_Bcrypt==1.0.1
flask_sqlalchemy==3.1.1
google_api_python_client==2.185.0
google_auth_oauthlib==1.2.2
protobuf==6.33.0
python-dotenv==1.2.1
pytz==2025.2
SQLAlchemy==2.0.44
Werkzeug==3.1.3
stripe==13.0.1
Flask-Migrate==4.1.0